requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
config==0.5.1
cryptography==45.0.5
Django==5.2.4
django-allauth==65.10.0
django-ckeditor==6.7.3
django-debug-toolbar==5.2.0
django-js-asset==3.1.2
django-json-widget==2.0.3
django-mptt==0.17.0
djangorestframework==3.16.0
idna==3.10
iniconfig==2.1.0
libsass==0.23.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
pytest==8.4.1
python-dotenv==1.1.1
pytils==0.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
transliterate==1.10.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
